requirements.txt
flask==3.0.0
groq==0.4.2
python-dotenv==1.0.1
pyyaml==6.0.1
deepdiff==7.0
jinja2==3.1.3
requests==2.31.0
